pyproject.toml
[project]
name = "jira-sentiment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.5",
    "httpx>=0.28.1",
    "jira>=3.10.5",
    "mcp>=1.16.0",
    "mcp-cli[cli]>=0.7.5",
    "python-dotenv>=1.0.1",
    "requests-oauthlib>=2.0.0",
    "sentence-transformers>=5.1.2",
]
